Fix select_window() by providing the session ID as argument to -t
Previosly select_window() would not work in nested tmux scenarios, as
there was no tmux session ID in the final tmux command. This is because
select_window() already adds the '-t' argument, hence the check in
cmd() does not add the session ID.

Thanks to Fabian Lesniak and Andreas Fried for pointing this out.

Fixes #161